pyproject.toml
[tool.poetry]
name = "backend-lab"
version = "0.1.0"
description = ""
authors = ["a1ekseevkiri11 <alekseevkirill30092004@mail.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
aiosqlite = "^0.20.0"
alembic = "^1.13.1"
bcrypt = "^4.1.3"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
cryptography = "^42.0.7"
pydantic-async-validation = "^0.1.4"
passlib = "^1.7.4"
py-bcrypt = "^0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
